+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 -   `<MobileStepper>` component for moving between pages.
 -   `<UserMenu>` component to display an avatar that opens up a bottomsheet menu.
 -   `leftComponent` prop to `<InfoListItem>`.
+-   Ability to use Avatar in `<Header>` actionItems.
 
 ### Changed
 -   Updated styles for the `<Drawer>`.